+++ src/main/java/bank/atm/dto/response/AccountResponseDto.java
@@ -1,5 +1,6 @@
 package bank.atm.dto.response;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -7,5 +8,5 @@ public class AccountResponseDto {
     private Long id;
     private Long accountNumber;
     private Long userId;
-    private long money;
+    private BigDecimal money;
 }
