+++ src/main/java/bank/atm/dto/response/BillResponseDto.java
@@ -1,9 +1,10 @@
 package bank.atm.dto.response;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class BillResponseDto {
     private Long id;
-    private long count;
+    private BigDecimal count;
 }
